Add an event handler to prevent form submission
Apparently returning false on the submit button event will *not* prevent
form submission in some browsers, such as Firefox. Adding an explicit
event handler to the form submission event and then using
`event.preventDefault()` should act as a more universal solution to
prevent the page from refreshing when we're doing an asynchronous
call to perform a same-page display of the results.